macos fix

+++ source_modules/kiwisdr_source/src/main.cpp
@@ -94,7 +94,7 @@ struct KiwiSDRSourceModule : public ModuleManager::Instance {
                 start = msg.substr(0, 3);
             }
             if (start == "MSG") {
-                flog::info("BIN/MSG: {} text: {}", msg.size(), msg);
+                flog::info("BIN/MSG: {} text: {}", (int64_t)msg.size(), msg);
             } else if (start == "SND") {
                 flog::info("{} Got sound: bytes={} )", (int64_t)currentTimeMillis(), (int64_t)msg.size());
                 long long int ctm = currentTimeMillis();
@@ -157,7 +157,7 @@ struct KiwiSDRSourceModule : public ModuleManager::Instance {
                         start += buf;
                     }
                 }
-                flog::info("BIN: {} bytes: {}", msg.size(), start);
+                flog::info("BIN: {} bytes: {}", (int64_t)msg.size(), start);
             }
         };
         lastPing = currentTimeMillis();
